Sample model corrected; Trainer updated with new model;

+++ 28_MachineLearning/Trial_01_LoadFromEnumerable/ML/Objects/FileInput.cs
@@ -8,16 +8,16 @@ namespace chapter03_logistic_regression.ML.Objects
         [LoadColumn(0)]
         public DateTime Date { get; set; }
         [LoadColumn(1)]
-        public decimal SmaDax { get; set; }
+        public float SmaDax { get; set; }
         [LoadColumn(2)]
-        public decimal SmaBrent { get; set; }
+        public float SmaBrent { get; set; }
         [LoadColumn(3)]
         public bool SmaDeltaDax { get; set; }
         [LoadColumn(4)]
         public bool SmaDeltaBrent { get; set; }
         [LoadColumn(5)]
-        public decimal CloseDax { get; set; }
+        public float CloseDax { get; set; }
         [LoadColumn(6)]
-        public decimal CloseBrent { get; set; }
+        public float CloseBrent { get; set; }
     }
 }
